bpf/tests: add scapy support
This commit adds some rudimentary scapy support in BPF unit tests
to simplify the process of writing and debugging unit tests.

Note: this commit does NOT replace pktgen.h, and tests written with
pktgen.h and scapy can coexist. In fact, scapy-based tests _still_ rely
on some pktgen.h infra.

The mechanics as of today are a bit hacky, and can be broken down in
two distinct stages:

* Compilation: before clang is invoked, `scapy/parser.py` goes
  through the C BPF unit test code looking for `BUF_DECL(NAME, ...)`
  invocations. As of today, `BUF_DECL()` only accepts a single
  argument which is the scapy packet definition, e.g.:

```
  BUF_DECL(ARP_REQ, arp_req)
```

And in `bpf/tests/scapy/pkt_defs.py`:

```
arp_req = Ether(dst=mac_bcast, src=mac_one)/ARP(op="who-has", psrc=v4_ext_one, pdst=v4_svc_one, hwsrc=mac_one, hwdst=mac_bcast));
```
  (but is kept as __VA_ARGS__ as it could accept more parameters to
  be passed to the parser, see TODO2)

  `scapy/parser.py` then creates a scapy packet, and generates
   a header file `scapy/output/gen_pkts.h` with the pkt byte array. This autogen
   header is then included via `scapy.h`. As of today, the buffer is allocated
   in stack. One can reference the scapy buffer - if reachable, in scope - using
   `SCAPY_BUF(NAME)`.

   At the same time, `scapy.h` defines a MACRO to compare a pkt (ctx)
   to a scapy buffer, to simplify writing pkt unit test assertions:
   `ASSERT_CTX_BUF_OFF(...)`.

   `scapy/pkt_defs.py` contains the packet definitions to be used in the tests,
   along with global constants (e.g. `mac_one`) reused in packet defintions.
   Packet definitions should be prepended with the test name or alike.
   Symbols in there are imported (via eval!) by `parser.py`, so they are
   available when `BUF_DECL()` is parsed, hugely simplifying buffer definitions.

* Output post-processing: BPF assertions are essentially the result of
  `memcmp()`, so while writing whole pkt checks is easier, debugging
  errors is difficult. To solve this, when `ASSERT_CTX_BUF_OFF()` fails,
  it uses `hexdump()` to dump both the expected and the actual packet.
  `scapy/trace_diff_pkts.py` in combination with the MAkefile then reads
  the packets from the `trace_pipe` and outputs the 'expected' and the 'got'
  using scapy's built-in dumper (`show()`).

TODO2: support no csum computation (don't invoke show2())
TODO3: _maybe_ scope buf names to the unit test itself (right now they are global in pkt_defs.py).
TODO4: don't use `trace_pipe` and post-processing to enrich assertions.
       It might be easier to use a map to store failed assertions, and
       post-process from there.

+++ bpf/lib/hexdump.h
@@ -49,8 +49,8 @@ void hexdump_len_off(const char *msg, const char *first_layer,
 	 * Further clamp dump len to either max of pkt or HD_MAX_BYTES
 	 * and make sure it doesn't get out of the pkt boundaries
 	 */
-	__u32 pkt_len = (data_end - data);
-	__u32 max_dump_len = len > HD_MAX_BYTES ? HD_MAX_BYTES : len;
+	__u16 pkt_len = (__u16)(data_end - data);
+	__u16 max_dump_len = len > HD_MAX_BYTES ? HD_MAX_BYTES : len;
 
 	max_dump_len = max_dump_len > pkt_len ? pkt_len : max_dump_len;
